folder structure reorganised. Had to change template path in app config.
